moved Token out into it's own header and implementation file
