+++ src/main.cpp
@@ -9,27 +9,10 @@
 #include <unordered_map>
 #include <utils/str_enum.hpp>
 #include <utils/str_format.hpp>
+#include <token.hpp>
 
 namespace tl {
 using number_t = double;
-
-STR_ENUM(TokenType,
-		 plus,	 // +
-		 minus,	 // -
-		 mult,	 // *
-		 div,	 // /
-		 mod,	 // %
-		 num,	 // [0-9]+(.[0-9]+)?
-		 string, // '[^']' | "[^"]"
-		 true_,	 // true
-		 false_, // false
-		 let,	 // let
-		 nil,	 // nil
-		 eof_,	 // '/0'
-		 error,	 // unknown
-		 eCOUNT_);
-
-using TT = TokenType;
 
 const static std::unordered_map<std::string_view, TokenType> keywords = {
 	// clang-format off
@@ -39,31 +22,6 @@ const static std::unordered_map<std::string_view, TokenType> keywords = {
 	{"false", TT::false_},
 	// clang-format on
 };
-
-struct TokenLocation {
-	std::size_t start = 0;
-	std::size_t length = 0;
-	std::size_t line = 0;
-	std::uint32_t column = 0;
-
-	std::string_view view(const std::string& source) const {
-		return std::string_view(source.c_str() + start, length);
-	}
-};
-
-struct Token {
-	TT type = TT::error;
-
-	TokenLocation location = {};
-
-	std::string_view raw(const std::string& source) const {
-		return location.view(source);
-	}
-};
-
-std::ostream& operator<<(std::ostream& os, const Token& token) {
-	return os << utils::toStr(token.type);
-}
 
 class Lexer {
   private:
